Make the model
