+++ backend/app/routes/__init__.py
@@ -1,0 +1,5 @@
+from .root import root
+
+routes = [
+    root,
+]
