+++ backend/app/routes/root.py
@@ -1,0 +1,5 @@
+from ..main import app
+
+@app.get("/")
+def root():
+    return {"message": "Hello World"}
